Replace the "stdout" with "stderr" in log stream
Currently the "logs" pages indicates that it is good practice to publish
logs unbuffered to a stream, and indicates that STDOUT is a good stream
to publish to.

However, the convention in UNIX systems is to publish diagnostic
information to STDERR, and publish program output to STDOUT:

> Under  normal  circumstances  every UNIX program has three streams
> opened for it when it starts up, one for input, one for output,
> and one for printing diagnostic or error messages.  These are
> typically attached to the user's terminal (see tty(4)) but might
> instead refer to files or other devices, depending on what the
> parent process chose to set up.  (See also the  "Redirection"
> section of sh(1).)
>
> The input stream is referred to as "standard input"; the output
> stream is referred to as "standard output"; and the error stream
> is referred to as "standard error".  These terms are abbreviated
> to form the symbols used to refer to these files, namely stdin,
> stdout, and stderr.

This pattern is fairly common in the standard linux tools, such as
`grep`:

  $ grep -ri 'fish' /sys > /dev/null
  grep: /sys/kernel/mm/page_idle/bitmap: Permission denied
  grep: /sys/kernel/slab/:d-0001024/remote_node_defrag_ratio: Permission denied

Where ">" sends `stdout` to `/dev/null` -- to nothing.